# server/src/main/resources/api/api.yaml
openapi: 3.0.3
info:
  title: Petka Infrastructure Project
  description: Api first project
  contact:
    name: Petka
    url: https://github.com/petka5/pis
    email: petar.georgiev@gmail.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
security:
  - bearerAuth: [ ]
tags:
  - name: Organization Pets
    description: Operations over the organization pets.
  - name: Operator Pets
    description: Operations over all pets.
paths:
  /operator/pets:
    get:
      tags:
        - Operator Pets
      description: Returns all pets from the system that the user has access to
      operationId: operatorFindPets
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
        - $ref: '#/components/parameters/sort'
        - $ref: '#/components/parameters/filter'
        - $ref: '#/components/parameters/includeDeleted'
        - $ref: '#/components/parameters/includeCount'
      responses:
        200:
          description: pet response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetPageResponse'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
    post:
      tags:
        - Operator Pets
      description: Creates a new pet in the store.  Duplicates are allowed
      operationId: operatorAddPet
      requestBody:
        description: Pet to add to the store
        content:
          application/json:
            schema:
                $ref: '#/components/schemas/OperatorPetRequest'
        required: true
      responses:
        201:
          description: pet response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetResponse'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
  /operator/pets/{id}:
    get:
      tags:
        - Operator Pets
      description: Returns resource by the id
      operationId: operatorFindPetById
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        200:
          description: pet response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetResponse'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
    delete:
      tags:
        - Operator Pets
      description: deletes a single pet based on the ID supplied
      operationId: operatorDeletePet
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        204:
          description: pet deleted
          content: { }
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
    patch:
      tags:
        - Operator Pets
      description: Partial update of the pet based on the ID supplied
      operationId: operatorUpdatePet
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        description: Pet elements that will be updated
        content:
          application/json:
            schema:
              type: object
        required: true
      responses:
        200:
          description: pet response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetResponse'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
  /{orgId}/pets:
    get:
      tags:
        - Organization Pets
      description: Returns all pets from the system that the user has access to
      operationId: orgFindPets
      parameters:
        - $ref: '#/components/parameters/orgId'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
        - $ref: '#/components/parameters/sort'
        - $ref: '#/components/parameters/filter'
        - $ref: '#/components/parameters/includeDeleted'
        - $ref: '#/components/parameters/includeCount'
      responses:
        200:
          description: pet response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetPageResponse'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
    post:
      tags:
        - Organization Pets
      description: Creates a new pet in the store.  Duplicates are allowed
      operationId: orgAddPet
      parameters:
        - $ref: '#/components/parameters/orgId'
      requestBody:
        description: Pet to add to the store
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PetRequest'
        required: true
      responses:
        201:
          description: pet response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetResponse'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
  /{orgId}/pets/{id}:
    get:
      tags:
        - Organization Pets
      description: Returns resource by the id
      operationId: orgFindPetById
      parameters:
        - $ref: '#/components/parameters/orgId'
        - $ref: '#/components/parameters/id'
      responses:
        200:
          description: pet response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetResponse'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
    delete:
      tags:
        - Organization Pets
      description: deletes a single pet based on the ID supplied
      operationId: orgDeletePet
      parameters:
        - $ref: '#/components/parameters/orgId'
        - $ref: '#/components/parameters/id'
      responses:
        204:
          description: pet deleted
          content: { }
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
    patch:
      tags:
        - Organization Pets
      description: Partial update of the pet based on the ID supplied
      operationId: orgUpdatePet
      parameters:
        - $ref: '#/components/parameters/orgId'
        - $ref: '#/components/parameters/id'
      requestBody:
        description: Pet elements that will be updated
        content:
          application/json:
            schema:
              type: object
        required: true
      responses:
        200:
          description: pet response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetResponse'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    id:
      name: id
      in: path
      description: ID of the resource
      required: true
      schema:
        type: string
        format: uuid
    includeDeleted:
      name: includeDeleted
      in: query
      description: By default deleted entities are not returned.
      required: false
      schema:
        type: boolean
        default: false
    includeCount:
      name: includeCount
      in: query
      description: Returns either slice or page. Default returns slice.
      required: false
      schema:
        type: boolean
        default: false
    page:
      name: page
      in: query
      description: Result page number.
      required: false
      schema:
        type: integer
        format: int32
        minimum: 0
        default: 0
    size:
      name: size
      in: query
      description: The size of the result page.
      required: false
      schema:
        type: integer
        format: int32
        default: 50
        maximum: 1000
    sort:
      name: sort
      in: query
      description: The name of a property to sort on and direction of sorting (asc or desc).
      required: false
      schema:
        type: string
    filter:
      name: filter
      in: query
      required: false
      allowEmptyValue: true
      schema:
        type: string
    orgId:
      name: orgId
      in: path
      description: Org ID
      required: true
      schema:
        example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        type: string
        format: uuid
  schemas:
    OperatorPetRequest:
      required:
        - orgId
      allOf:
        - $ref: '#/components/schemas/PetRequest'
      properties:
        orgId:
          type: string
          format: uuid
    PetRequest:
      required:
        - name
        - kind
        - age
        - type
      type: object
      properties:
        name:
          type: string
        kind:
          type: string
        age:
          type: integer
          format: int64
        type:
          $ref: '#/components/schemas/Type'
    PetResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        orgId:
          type: string
          format: uuid
        name:
          type: string
        kind:
          type: string
        age:
          type: integer
          format: int64
        type:
          $ref: '#/components/schemas/Type'
        version:
          type: integer
          format: int64
        createDateTime:
          type: string
        updateDateTime:
          type: string
        deleted:
          type: boolean
    PetPageResponse:
      allOf:
        - $ref: '#/components/schemas/PageResponse'
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/PetResponse'
    PageResponse:
      type: object
      description: Response for Get All
      properties:
        totalElements:
          type: integer
        totalPages:
          type: integer
        numberOfElements:
          type: integer
        hasNext:
          type: boolean
        number:
          type: integer
        size:
          type: integer
    Type:
      type: string
      enum:
        - DOMESTIC
        - WILD
    errorModel:
      required:
        - code
        - message
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
x-api-first-error-mapping:
  404:
    - java.util.NoSuchElementException
    - java.lang.IndexOutOfBoundsException
  406:
    - java.lang.UnsupportedOperationException
    - java.lang.IllegalArgumentException
    - java.lang.IllegalStateException
  409:
    - java.util.ConcurrentModificationException
    - java.util.EmptyStackException
  500:
    - java.sql.SQLException